Finished refactoring.
