+++ backend/jobapplicationtrackerbackend/src/main/java/com/jpettit/jobapplicationtrackerbackend/enums/JobApplicationURLS.java
@@ -1,4 +1,4 @@
-package com.jpettit.jobapplicationtrackerbackend.helpers;
+package com.jpettit.jobapplicationtrackerbackend.enums;
 
 public class JobApplicationURLS {
     public static final String getJobAppCard = "/api/jobAppCard/{sessionId}";
