+++ backend/jobapplicationtrackerbackend/src/main/java/com/jpettit/jobapplicationtrackerbackend/enums/ProjectEnvironment.java
@@ -1,0 +1,5 @@
+package com.jpettit.jobapplicationtrackerbackend.enums;
+
+public enum ProjectEnvironment {
+    TEST, DEV, PROD
+}
