--- backend/jobapplicationtrackerbackend/src/main/java/com/jpettit/jobapplicationtrackerbackend/helpers/ProjectEnvironment.java
@@ -1,5 +1,0 @@
-package com.jpettit.jobapplicationtrackerbackend.helpers;
-
-public enum ProjectEnvironment {
-    TEST, DEV, PROD
-}
