+++ backend/jobapplicationtrackerbackend/src/main/java/com/jpettit/jobapplicationtrackerbackend/helpers/ProjectEnvironmentReader.java
@@ -1,5 +1,7 @@
 package com.jpettit.jobapplicationtrackerbackend.helpers;
 
+import com.jpettit.jobapplicationtrackerbackend.enums.ProjectEnvironment;
+
 import java.util.HashMap;
 
 public class ProjectEnvironmentReader {
