+++ backend/jobapplicationtrackerbackend/src/main/java/com/jpettit/jobapplicationtrackerbackend/helpers/Querier.java
@@ -1,5 +1,7 @@
 package com.jpettit.jobapplicationtrackerbackend.helpers;
 
+import com.jpettit.jobapplicationtrackerbackend.enums.ProjectEnvironment;
+
 public abstract class Querier {
     abstract String getTableName(ProjectEnvironment env);
 }
