+++ backend/jobapplicationtrackerbackend/src/test/java/com/jpettit/jobapplicationtrackerbackend/controllers/JobApplicationControllerTest.java
@@ -4,14 +4,13 @@ import com.jpettit.jobapplicationtrackerbackend.daos.JobApplicationDAO;
 import com.jpettit.jobapplicationtrackerbackend.daos.UserDAO;
 import com.jpettit.jobapplicationtrackerbackend.enums.AppProperties;
 import com.jpettit.jobapplicationtrackerbackend.helpers.JobApplicationControllerTestHelper;
-import com.jpettit.jobapplicationtrackerbackend.helpers.JobApplicationURLS;
+import com.jpettit.jobapplicationtrackerbackend.enums.JobApplicationURLS;
 import com.jpettit.jobapplicationtrackerbackend.models.JobApplicationCard;
 import com.jpettit.jobapplicationtrackerbackend.models.ResultPair;
 import com.jpettit.jobapplicationtrackerbackend.models.Session;
 import com.jpettit.jobapplicationtrackerbackend.models.User;
 import com.jpettit.jobapplicationtrackerbackend.services.JobApplicationService;
 import com.jpettit.jobapplicationtrackerbackend.services.UserService;
-import org.json.JSONObject;
 import org.junit.jupiter.api.*;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
