+++ backend/jobapplicationtrackerbackend/src/test/java/com/jpettit/jobapplicationtrackerbackend/helpers/JobAppQuerierTest.java
@@ -1,11 +1,12 @@
 package com.jpettit.jobapplicationtrackerbackend.helpers;
 
 import com.jpettit.jobapplicationtrackerbackend.models.ResultPair;
-import testhelpers.JobAppQuerierTestHelpers;
+import testhelpers.helpers.JobAppQuerierTestHelpers;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 
-import static testhelpers.JobAppQuerierTestHelpers.*;
+import static testhelpers.helpers.JobAppQuerierTestHelpers.*;
+import static testhelpers.helpervars.JobAppQuerierTestHelperVars.*;
 
 class JobAppQuerierTest {
     private JobAppQuerier sut;
@@ -17,27 +18,27 @@ class JobAppQuerierTest {
 
     @Test
     public void testGetAllJobAppCards_whenPassedInTestEnvAndUser1_shouldReturnQueryWithUser1AndDevEnv() {
-        sut = JobAppQuerierTestHelpers.createJobAppQuerier(TEST_ENV);
+        sut = createJobAppQuerier(TEST_ENV);
 
         final String EXPECTEDQUERY = getQuery(TEST_ENV, "", USER1);
         final ResultPair<String> EXPECTED_RESULT = new ResultPair<>(EXPECTEDQUERY, "");
 
         final ResultPair<String> ACTUAL_RESULT = sut.getAllJobAppCards(USER1);
 
-        JobAppQuerierTestHelpers.comparePairs(ACTUAL_RESULT, EXPECTED_RESULT);
+        comparePairs(ACTUAL_RESULT, EXPECTED_RESULT);
     }
 
     @Test
     public void testGetAllJobAppCards_whenPassedInDevEnvAndUser1_shouldReturnQueryWithUser1AndDevEnv() {
         final String USER1 = "user1";
-        sut = JobAppQuerierTestHelpers.createJobAppQuerier(DEV_ENV);
+        sut = createJobAppQuerier(DEV_ENV);
 
-        final String EXPECTEDQUERY = JobAppQuerierTestHelpers.getQuery(DEV_ENV, "", USER1);
+        final String EXPECTEDQUERY = getQuery(DEV_ENV, "", USER1);
         final ResultPair<String> EXPECTED_RESULT = new ResultPair<>(EXPECTEDQUERY, "");
 
         final ResultPair<String> ACTUAL_RESULT = sut.getAllJobAppCards(USER1);
 
-        JobAppQuerierTestHelpers.comparePairs(ACTUAL_RESULT, EXPECTED_RESULT);
+        comparePairs(ACTUAL_RESULT, EXPECTED_RESULT);
     }
 
     @Test
