+++ backend/jobapplicationtrackerbackend/src/test/java/com/jpettit/jobapplicationtrackerbackend/models/JobApplicationCardTest.java
@@ -1,12 +1,17 @@
 package com.jpettit.jobapplicationtrackerbackend.models;
 
 import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Test;
+import testhelpers.helpers.JobApplicationCardTestHelpers;
+import testhelpers.helpervars.JobApplicationCardHelperVars;
 
 class JobApplicationCardTest {
-    JobApplicationCard sut;
 
-    @AfterEach
-    void tearDown() {
-        sut = null;
+    @Test
+    public void testCreateFromJobApp_whenInitedWithJobApp_gettersShouldReturnGivenJobAppValues() {
+        final JobApplication JOB_APP = JobApplicationCardHelperVars.createDefaultJobApp();
+        final JobApplicationCard CARD = JobApplicationCard.createFromJobApp(JOB_APP);
+
+        JobApplicationCardTestHelpers.verifyIfJobAppCardAndJobAppAreEqual(CARD, JOB_APP);
     }
 }
