+++ backend/jobapplicationtrackerbackend/src/test/java/com/jpettit/jobapplicationtrackerbackend/services/UserServiceTest.java
@@ -3,7 +3,7 @@ package com.jpettit.jobapplicationtrackerbackend.services;
 import com.jpettit.jobapplicationtrackerbackend.daos.UserDAO;
 import com.jpettit.jobapplicationtrackerbackend.helpers.PasswordEncoder;
 import com.jpettit.jobapplicationtrackerbackend.models.*;
-import testhelpers.UserServiceHelper;
+import testhelpers.helpers.UserServiceHelper;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -15,6 +15,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import org.mockito.junit.jupiter.MockitoSettings;
 import org.mockito.quality.Strictness;
+import testhelpers.helpervars.UserServiceTestHelperVars;
 
 import java.time.LocalDate;
 import java.util.Optional;
@@ -31,11 +32,6 @@ class UserServiceTest {
     @Mock
     private PasswordEncoder encoder;
 
-//    @BeforeEach
-//    void setUp() {
-//
-//    }
-
     @AfterAll
     private static void tearDown() {
         sut = null;
@@ -43,7 +39,7 @@ class UserServiceTest {
 
     @Test
     public void testCreateUser_whenPassedInNonExistingUser_shouldReturnSessionIdAndNoErrorMessage() {
-        final Optional<User> user = Optional.of(UserServiceHelper.user);
+        final Optional<User> user = Optional.of(UserServiceTestHelperVars.user);
         final Session session = user.get().getSession();
         final String USER_PASSWORD = user.get().getPassword();
 
@@ -58,7 +54,7 @@ class UserServiceTest {
 
     @Test
     public void testCreateUser_whenPassedInExistingUser_shouldReturnEmptyValueAndUserExists() {
-        final Optional<User> user = Optional.of(UserServiceHelper.user);
+        final Optional<User> user = Optional.of(UserServiceTestHelperVars.user);
         final String EXPECTED_ERROR_MSG = "User exists";
 
         Mockito.when(userDAO.getByUsername(user.get().getUsername())).thenReturn(user);
@@ -71,7 +67,7 @@ class UserServiceTest {
 
     @Test
     public void testValidateUserLogin_whenPassedInNonExistingUser_shouldReturnEmptyValueAndErrorMessage() {
-        final Login LOGIN = UserServiceHelper.nonExistantLogin;
+        final Login LOGIN = UserServiceTestHelperVars.nonExistantLogin;
         final ResultPair<String> pair = new ResultPair<>("", "Username or password doesn't exist");
         Mockito.when(userDAO.getPasswordForUser(LOGIN)).thenReturn(pair);
 
@@ -83,7 +79,7 @@ class UserServiceTest {
 
     @Test
     public void testValidateUserLogin_whenPassedInDifferentPasswordForUser_shouldReturnEmptyValueAndErrorMessage() {
-       final Login LOGIN = UserServiceHelper.login;
+       final Login LOGIN = UserServiceTestHelperVars.login;
 
        Mockito.when(userDAO.getPasswordForUser(LOGIN)).thenReturn(new ResultPair<>("manWord", ""));
 
@@ -96,7 +92,7 @@ class UserServiceTest {
 
     @Test
     public void testValidateUserLogin_whenPassedInExistingLogin_shouldReturnSessionNameAndNoErrorMessage() {
-        final Login LOGIN = UserServiceHelper.login;
+        final Login LOGIN = UserServiceTestHelperVars.login;
 
         Mockito.when(userDAO.getPasswordForUser(LOGIN))
                 .thenReturn(new ResultPair<>(LOGIN.getPassword(), ""));
@@ -112,7 +108,7 @@ class UserServiceTest {
 
     @Test
     public void testValidateUserLogin_whenPassedInLoginAndCantUpdateSession_shouldReturnNoValueAndCouldNotUpdateSession() {
-        final Login LOGIN = UserServiceHelper.login;
+        final Login LOGIN = UserServiceTestHelperVars.login;
 
         Mockito.when(userDAO.getPasswordForUser(LOGIN))
                 .thenReturn(new ResultPair<>(LOGIN.getPassword(), ""));
@@ -130,7 +126,7 @@ class UserServiceTest {
 
     @Test
     public void testGetUsernameBySessionId_whenPassedInValidSessionId_shouldReturnAssociatedUsernameAndNoErrorMessage() {
-        final User USER = UserServiceHelper.user;
+        final User USER = UserServiceTestHelperVars.user;
         final String SESSION_ID = USER.getSession().getSessionName();
         final ResultPair<String> EXPECTED_PAIR = new ResultPair<>(USER.getUsername(), "");
 
@@ -143,7 +139,7 @@ class UserServiceTest {
 
     @Test
     public void testGetUsernameBySessionId_whenPassedEmptySessionId_shouldReturnNoUsernameAndEmptySessionIdErrorMessage() {
-        final User USER = UserServiceHelper.user;
+        final User USER = UserServiceTestHelperVars.user;
         final String SESSION_ID = "";
         final ResultPair<String> EXPECTED_PAIR = new ResultPair<>("", userDAO.EMPTY_SESSIONID);
 
@@ -156,7 +152,7 @@ class UserServiceTest {
 
     @Test
     public void testHasSessionExpired_whenPassedInAccessDateOfDec201999AndSessionExpDateIsJan12000_shouldReturnFalse() {
-        final Session SESSION = UserServiceHelper.user.getSession();
+        final Session SESSION = UserServiceTestHelperVars.user.getSession();
         final ResultPair<Optional<LocalDate>> EXPECTED_PAIR = new ResultPair<>(Optional.of(SESSION.getExpirationDate()), "");
         final LocalDate DEC20_1999 = LocalDate.of(1999, 12, 20);
 
@@ -169,7 +165,7 @@ class UserServiceTest {
 
     @Test
     public void testHasSessionExpired_whenPassedInAccessDateOfJan2_2000AndSessionExpDateIsJan1_2000_shouldReturnFalse() {
-        final Session SESSION = UserServiceHelper.user.getSession();
+        final Session SESSION = UserServiceTestHelperVars.user.getSession();
         final ResultPair<Optional<LocalDate>> EXPECTED_PAIR = new ResultPair<>(Optional.of(SESSION.getExpirationDate()), "");
         final LocalDate JAN_2_2000 = LocalDate.of(2000, 1, 2);
 
@@ -183,7 +179,7 @@ class UserServiceTest {
     @Test
     public void testHasSessionExpired_whenSessionCantBeFound_shouldReturnTrueAndCantFindSessionWithThatId() {
         final ResultPair<Optional<LocalDate>> INVALID_PAIR = new ResultPair<>(Optional.empty(), sut.SESSION_NOT_FOUND);
-        final Session SESSION = UserServiceHelper.user.getSession();
+        final Session SESSION = UserServiceTestHelperVars.user.getSession();
 
         Mockito.when(userDAO.getSessionExpDateBySessionId(Mockito.anyString())).thenReturn(INVALID_PAIR);
