+++ backend/jobapplicationtrackerbackend/src/test/java/testhelpers/helpers/JobApplicationCardTestHelpers.java
@@ -1,4 +1,4 @@
-package testhelpers;
+package testhelpers.helpers;
 
 import com.jpettit.jobapplicationtrackerbackend.models.Interview;
 import com.jpettit.jobapplicationtrackerbackend.models.JobApplication;
@@ -16,26 +16,6 @@ public class JobApplicationCardTestHelpers {
     private static final String COMPANY_KEY = "company";
     private static final String COUNT_KEY = "count";
     private static final String ID_KEY = "id";
-
-    public static JobApplication createDefaultJobApp() {
-        final ArrayList<Interview> INTERVIEWS = createDefaultInterviews();
-
-        return JobApplication.createJobApplication("iOS Developer", "EJ", "1",
-                LocalDate.now(), "user1", "Pending", INTERVIEWS);
-    }
-
-    public static ArrayList<Interview> createDefaultInterviews() {
-        final Interview[] INTERVIEWS = {
-                Interview.createInterview("1", "1", "Phone interview",
-                        Date.valueOf(LocalDate.now()), "Next Round"),
-                Interview.createInterview("2", "1", "Phone interview",
-                        Date.valueOf(LocalDate.now()), "Next Round"),
-                Interview.createInterview("3", "1", "Phone interview",
-                        Date.valueOf(LocalDate.now()), "Next Round")
-        };
-
-        return new ArrayList<>(Arrays.asList(INTERVIEWS));
-    }
 
     public static void verifyIfJobAppCardAndJobAppAreEqual(JobApplicationCard card, JobApplication jobApp) {
         HashMap<String, String> map = getErrorMessages(card, jobApp);
