+++ backend/jobapplicationtrackerbackend/src/test/java/testhelpers/helpers/TestPair.java
@@ -1,4 +1,4 @@
-package testhelpers;
+package testhelpers.helpers;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
