+++ backend/jobapplicationtrackerbackend/src/test/java/testhelpers/helpers/UserServiceHelper.java
@@ -1,7 +1,8 @@
-package testhelpers;
+package testhelpers.helpers;
 
 import com.jpettit.jobapplicationtrackerbackend.models.*;
 import org.junit.jupiter.api.Assertions;
+import testhelpers.helpers.TestPair;
 
 import java.time.LocalDate;
 import java.time.format.TextStyle;
@@ -9,16 +10,6 @@ import java.util.Locale;
 import java.util.Optional;
 
 public class UserServiceHelper {
-    public static final User user = User.createUser("u1", "e1",
-            "p1",
-            Session.createSession("invalidSession", LocalDate.of(2000, 1, 1)),
-            0L);
-    public static final User user2 = User.createUser("u2", "e2", "p2",
-            Session.createSession("goodSession", LocalDate.of(2000, 1, 1)), 1L);
-    public static final Optional<User> nullUser = Optional.empty();
-    public static final Login login = Login.createLogin("u1", "p1");
-    public static final Login nonExistantLogin = Login.createLogin("dude", "p5");
-
     public static void verifyUserServiceUserPair(UserServiceIntPair actual, UserServiceIntPair expected) {
         TestPair<Integer> countPair = new TestPair<>(actual.getValue(), expected.getValue());
         TestPair<String> messagePair = new TestPair<>(actual.getMessage(), expected.getMessage());
