+++ backend/jobapplicationtrackerbackend/src/test/java/testhelpers/helpervars/JobApplicationDAOTestHelperVars.java
@@ -1,4 +1,4 @@
-package testhelpers;
+package testhelpers.helpervars;
 
 import com.jpettit.jobapplicationtrackerbackend.models.JobApplicationCard;
