+++ backend/jobapplicationtrackerbackend/src/test/java/testhelpers/helpervars/PasswordEncoderTestHelperVars.java
@@ -1,4 +1,4 @@
-package testhelpers;
+package testhelpers.helpervars;
 
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
