+++ backend/jobapplicationtrackerbackend/src/test/java/testhelpers/helpervars/UserControllerTestHelperVars.java
@@ -1,4 +1,4 @@
-package testhelpers;
+package testhelpers.helpervars;
 
 import com.jpettit.jobapplicationtrackerbackend.models.ResultPair;
 import com.jpettit.jobapplicationtrackerbackend.models.Session;
